+++ Assets/Scripts/GameSettings/SettingGroups/RulesSettingsControls.cs
@@ -14,10 +14,10 @@ namespace Blockstacker.GameSettings.SettingGroups
         public RotationSystemType RotationSystem = RotationSystemType.SRS;
         public string CustomRotationSystemName = "";
         public RotationSystem ActiveRotationSystem;
-        public float PiecePlacedDelay = 0;
-        public float LineClearDelay = 0;
+        public double PiecePlacedDelay = 0;
+        public double LineClearDelay = 0;
         public OnTouchGround OnTouchGround = OnTouchGround.LimitedTime;
-        public float OnTouchGroundAmount = 5;
+        public double OnTouchGroundAmount = 5;
         public bool ShowGhostPiece = true;
         public bool OverrideHandling = false;
         public HandlingSettings Handling = new();
