+++ Assets/Scripts/GameSettings/SettingGroups/RulesSettingsLevelling.cs
@@ -6,8 +6,8 @@ namespace Blockstacker.GameSettings.SettingGroups
     [Serializable]
     public class RulesSettingsLevelling
     {
-        public float Gravity = .02f;
-        public float LockDelay = .5f;
+        public double Gravity = .02f;
+        public double LockDelay = .5f;
         public uint StartingLevel = 1;
         public LevellingSystem LevellingSystem = LevellingSystem.None;
         public string CustomLevellingScriptName = "";
